adds a guard clause to delete-article.php to make sure I can only delete an article using the post method and not using the get method

+++ delete-article.php
@@ -21,21 +21,24 @@ if (isset($_GET['id'])) {
   die("id not supplied, article not found");
 }
 
-$sql = "DELETE FROM article WHERE id = ?";
+if ($_SERVER["REQUEST_METHOD"] === "POST") {
+
+  $sql = "DELETE FROM article WHERE id = ?";
 
 
-$stmt = mysqli_prepare($conn, $sql);
+  $stmt = mysqli_prepare($conn, $sql);
 
 
-if ($stmt === false) {
-  echo mysqli_error($conn);
-} else {
-  mysqli_stmt_bind_param($stmt, "i", $id);
-
-
-  if (mysqli_stmt_execute($stmt)) {
-    redirect("/cms/index.php");
+  if ($stmt === false) {
+    echo mysqli_error($conn);
   } else {
-    echo mysqli_stmt_error($stmt);
+    mysqli_stmt_bind_param($stmt, "i", $id);
+
+
+    if (mysqli_stmt_execute($stmt)) {
+      redirect("/cms/index.php");
+    } else {
+      echo mysqli_stmt_error($stmt);
+    }
   }
 }
